MOBILE-1047 Handle error statusCode as a Int

+++ apio-consumer/src/main/kotlin/com/liferay/apio/consumer/util/RequestUtil.kt
@@ -90,7 +90,7 @@ internal class RequestUtil {
 				ThingParser.stringToJsonMap(it)
 			}?.let {
 				RequestFailedException(
-					it["statusCode"] as Number,
+					(it["statusCode"] as Number).toInt(),
 					ThingParser.parseType(it["@type"]),
 					it["title"] as String,
 					it["description"] as? String
